whatsapp-loyalty-bot/requirements.txt
Flask==2.3.3
firebase-admin==6.3.0
google-cloud-vision==3.4.2
python-dotenv==1.0.1
twilio==9.0.4
